[INJIMOB-3558]: Moving replacement helper methods to Json Pointer Resolver class
