+++ kotlin/injivcrenderer/src/commonMain/kotlin/io/mosip/injivcrenderer/templateEngine/svg/JsonPointerResolver.kt
@@ -2,12 +2,65 @@ package io.mosip.injivcrenderer.templateEngine.svg
 
 import com.fasterxml.jackson.core.JsonPointer
 import com.fasterxml.jackson.databind.JsonNode
+import io.mosip.injivcrenderer.constants.Constants.QR_CODE_FALLBACK_IMAGE_ID
+import io.mosip.injivcrenderer.constants.Constants.QR_CODE_IMAGE_ID
+import io.mosip.injivcrenderer.constants.Constants.QR_CODE_PLACEHOLDER
+import io.mosip.injivcrenderer.constants.Constants.QR_IMAGE_PREFIX
+import io.mosip.injivcrenderer.constants.Constants.RENDER_PROPERTY
+import io.mosip.injivcrenderer.constants.Constants.TEMPLATE
 import io.mosip.injivcrenderer.constants.VcRendererErrorCodes.MISSING_JSON_PATH
+import io.mosip.injivcrenderer.qrCode.QrCodeGenerator
+import io.mosip.injivcrenderer.qrCode.QrCodeGenerator.Companion.DEFAULT_FALLBACK_QR_BASE64
 import java.util.logging.Level
 import java.util.logging.Logger
 
 class JsonPointerResolver(private val traceabilityId: String) {
     private val className = JsonPointerResolver::class.simpleName
+
+    fun replaceSvgPlaceholders(
+        svg: String,
+        vcJsonNode: JsonNode,
+        renderMethodElement: JsonNode,
+        vcJsonString: String
+    ): String {
+        val svgWithQrCodeReplaced = replaceQrCodePlaceholder(svg, vcJsonString)
+        return replaceVcPlaceholders(svgWithQrCodeReplaced, vcJsonNode, renderMethodElement)
+    }
+
+    private fun replaceVcPlaceholders(svg: String, vcJsonNode: JsonNode, element: JsonNode): String {
+        val renderProperties =
+            element.path(TEMPLATE).path(RENDER_PROPERTY)
+                .takeIf { it.isArray }
+                ?.map { it.asText() }
+
+        return JsonPointerResolver(traceabilityId).replacePlaceholders(
+            svgTemplate = svg,
+            jsonNode = vcJsonNode,
+            renderProperties = renderProperties
+        )
+    }
+
+    private fun replaceQrCodePlaceholder(svg: String, vcJsonString: String): String {
+        return if (!svg.contains(QR_CODE_PLACEHOLDER)) {
+            svg
+        } else {
+            val qrBase64 = try {
+                QrCodeGenerator(traceabilityId).generateQRCodeImage(vcJsonString)
+            } catch (e: Exception) {
+                println("[$traceabilityId] QR generation failed: ${e.message}")
+                null
+            }
+
+            val finalQrBase64 = qrBase64.takeUnless { it.isNullOrEmpty() } ?: DEFAULT_FALLBACK_QR_BASE64
+            val qrImageTag = "$QR_IMAGE_PREFIX,$finalQrBase64"
+
+            val imageId = if (qrBase64.isNullOrEmpty()) QR_CODE_FALLBACK_IMAGE_ID else QR_CODE_IMAGE_ID
+
+            return svg
+                .replace(QR_CODE_PLACEHOLDER, qrImageTag)
+                .replace(QR_CODE_IMAGE_ID, imageId)
+        }
+    }
 
     /**
      * Replaces placeholders in an SVG template using a Verifiable Credential JSON for values.
